requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
langchain-ollama
chromadb
pypdf
